--- src/app/app-routing.module.ts
@@ -1,19 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { WelcomeComponent } from './pages/welcome/welcome.component';
-
-const routes: Routes = [
-  { path: 'welcome', component: WelcomeComponent },
-  {
-    path: 'league',
-    loadChildren: () =>
-      import('./pages/league/league.module').then((m) => m.LeagueModule),
-  },
-  { path: '**', redirectTo: 'welcome', pathMatch: 'prefix' },
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule],
-})
-export class AppRoutingModule {}
