--- src/app/app.module.ts
@@ -1,47 +1,0 @@
-import { NgModule, isDevMode } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import { provideHttpClient } from '@angular/common/http';
-import { ReactiveFormsModule } from '@angular/forms';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { WelcomeComponent } from './pages/welcome/welcome.component';
-import { StoreModule } from '@ngrx/store';
-import { EffectsModule } from '@ngrx/effects';
-import { leagueReducer } from './store/league/league.reducer';
-import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-import { rostersReducer } from './store/rosters/rosters.reducers';
-import { GlobalEffects } from './store/global.effects';
-import { managersReducer } from './store/managers/managers.reducers';
-import { sharedDataReducer } from './store/global.reducers';
-import { ServiceWorkerModule } from '@angular/service-worker';
-import { InputTextModule } from 'primeng/inputtext';
-import { ButtonModule } from 'primeng/button';
-@NgModule({
-  declarations: [AppComponent, WelcomeComponent],
-  bootstrap: [AppComponent],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    BrowserAnimationsModule,
-    ReactiveFormsModule,
-    InputTextModule,
-    ButtonModule,
-    StoreModule.forRoot({
-      sharedData: sharedDataReducer,
-      leagueData: leagueReducer,
-      rosterData: rostersReducer,
-      managersData: managersReducer,
-    }),
-    EffectsModule.forRoot([GlobalEffects]),
-    StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: !isDevMode(), connectInZone: true }),
-    ServiceWorkerModule.register('ngsw-worker.js', {
-        enabled: !isDevMode(),
-        // Register the ServiceWorker as soon as the application is stable
-        // or after 30 seconds (whichever comes first).
-        registrationStrategy: 'registerWhenStable:30000'
-    })
-  ],
-  providers: [provideHttpClient()] })
-export class AppModule {}
