+++ src/app/pages/league/graph/graph.component.ts
@@ -3,14 +3,18 @@ import { Store } from '@ngrx/store';
 import { Subscription, tap } from 'rxjs';
 import { LeaguePageData } from 'src/app/interfaces/leaguePageData';
 import { selectLeaguePageData } from 'src/app/store/selectors';
+import { LeaguePageHeaderComponent } from '../../../components/league-page-header/league-page-header.component';
+import { CommonModule } from '@angular/common';
 
 const SUBTITLE_TEXT =
   'Better teams are further right and more successful teams are further up \n Better managers have larger dots - dot size is proportional to the percentage of max points the team has scored';
 
 @Component({
-  selector: 'app-graph',
-  templateUrl: './graph.component.html',
-  styleUrls: ['./graph.component.scss'],
+    selector: 'app-graph',
+    templateUrl: './graph.component.html',
+    styleUrls: ['./graph.component.scss'],
+    standalone: true,
+    imports: [CommonModule, LeaguePageHeaderComponent],
 })
 export class GraphComponent implements OnInit, OnDestroy {
   readonly #store = inject(Store);
