+++ src/app/pages/welcome/welcome.component.ts
@@ -4,13 +4,21 @@ import { Store } from '@ngrx/store';
 import { leagueEntryRequest } from '../../store/global.actions';
 import { SleeperApiService } from '../../api/sleeper-api.service';
 import { League } from '../../interfaces/league';
-import { FormControl, Validators } from '@angular/forms';
+import { FormControl, Validators, ReactiveFormsModule } from '@angular/forms';
 import { catchError, take, tap, throwError } from 'rxjs';
+import { InputTextModule } from 'primeng/inputtext';
+import { ButtonModule } from 'primeng/button';
 
 @Component({
-  selector: 'app-welcome',
-  templateUrl: './welcome.component.html',
-  styleUrls: ['./welcome.component.scss'],
+    selector: 'app-welcome',
+    templateUrl: './welcome.component.html',
+    styleUrls: ['./welcome.component.scss'],
+    standalone: true,
+    imports: [
+        ReactiveFormsModule,
+        InputTextModule,
+        ButtonModule,
+    ],
 })
 export class WelcomeComponent {
   leagueInputForm = new FormControl<string>('', [Validators.required, Validators.minLength(3)]);
